FIX: Serialization issues for CLI
